+++ modules/distributor/forwarder.go
@@ -13,6 +13,7 @@ import (
 	"go.uber.org/multierr"
 
 	"github.com/grafana/tempo/modules/distributor/queue"
+	"github.com/grafana/tempo/modules/generator"
 	"github.com/grafana/tempo/modules/overrides"
 )
 
@@ -34,12 +35,13 @@ var (
 	}, []string{"tenant"})
 )
 
-type forwardFunc func(ctx context.Context, tenantID string, keys []uint32, traces []*rebatchedTrace) error
+type forwardFunc func(ctx context.Context, tenantID string, keys []uint32, traces []*rebatchedTrace, noGenerateMetrics bool) error
 
 type request struct {
-	tenantID string
-	keys     []uint32
-	traces   []*rebatchedTrace
+	tenantID          string
+	keys              []uint32
+	traces            []*rebatchedTrace
+	noGenerateMetrics bool
 }
 
 // generatorForwarder queues up traces to be sent to the metrics-generators
@@ -84,7 +86,7 @@ func (f *generatorForwarder) SendTraces(ctx context.Context, tenantID string, ke
 	}
 
 	q := f.getOrCreateQueue(tenantID)
-	err := q.Push(ctx, &request{tenantID: tenantID, keys: keys, traces: traces})
+	err := q.Push(ctx, &request{tenantID: tenantID, keys: keys, traces: traces, noGenerateMetrics: generator.ExtractNoGenerateMetrics(ctx)})
 	if err != nil {
 		_ = level.Error(f.logger).Log("msg", "failed to push traces to queue", "tenant", tenantID, "err", err)
 		metricForwarderPushesFailures.WithLabelValues(tenantID).Inc()
@@ -216,7 +218,7 @@ func (f *generatorForwarder) stop(_ error) error {
 }
 
 func (f *generatorForwarder) processFunc(ctx context.Context, data *request) {
-	if err := f.forwardFunc(ctx, data.tenantID, data.keys, data.traces); err != nil {
+	if err := f.forwardFunc(ctx, data.tenantID, data.keys, data.traces, data.noGenerateMetrics); err != nil {
 		_ = level.Warn(f.logger).Log("msg", "failed to forward request to metrics generator", "err", err)
 	}
 }