+++ modules/frontend/search_handlers_test.go
@@ -587,7 +587,7 @@ func TestSearchAccessesCache(t *testing.T) {
 	hash := hashForSearchRequest(&tempopb.SearchRequest{Query: query, Limit: 3, SpansPerSpanSet: 2})
 	start := uint32(10)
 	end := uint32(20)
-	cacheKey := searchJobCacheKey(tenant, hash, int64(start), int64(end), meta, 0, 1)
+	cacheKey := searchJobCacheKey(tenant, hash, time.Unix(int64(start), 0), time.Unix(int64(end), 0), meta, 0, 1)
 
 	// confirm cache key coesn't exist
 	_, bufs, _ := c.Fetch(context.Background(), []string{cacheKey})
@@ -718,7 +718,7 @@ func BenchmarkSearchPipeline(b *testing.B) {
 // frontendWithSettings returns a new frontend with the given settings. any nil options
 // are given "happy path" defaults
 func frontendWithSettings(t require.TestingT, next pipeline.RoundTripper, rdr tempodb.Reader, cfg *Config, cacheProvider cache.Provider,
-	opts ...func(*Config),
+	opts ...func(*Config, *overrides.Config),
 ) *QueryFrontend {
 	if next == nil {
 		next = &mockRoundTripper{
@@ -802,11 +802,13 @@ func frontendWithSettings(t require.TestingT, next pipeline.RoundTripper, rdr te
 		}
 	}
 
+	overridesCfg := &overrides.Config{}
+
 	for _, o := range opts {
-		o(cfg)
+		o(cfg, overridesCfg)
 	}
 
-	o, err := overrides.NewOverrides(overrides.Config{}, nil, prometheus.DefaultRegisterer)
+	o, err := overrides.NewOverrides(*overridesCfg, nil, prometheus.DefaultRegisterer)
 	require.NoError(t, err)
 
 	f, err := New(*cfg, next, o, rdr, cacheProvider, "", log.NewNopLogger(), nil)