+++ modules/frontend/tag_handlers_test.go
@@ -10,6 +10,7 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"regexp"
+	"sort"
 	"testing"
 	"time"
 
@@ -18,7 +19,10 @@ import (
 	"github.com/gogo/status"
 	"github.com/gorilla/mux"
 	"github.com/grafana/dskit/user"
+	"github.com/grafana/tempo/modules/overrides"
+	"github.com/grafana/tempo/pkg/api"
 	"github.com/grafana/tempo/pkg/cache"
+	"github.com/grafana/tempo/pkg/search"
 	"github.com/grafana/tempo/pkg/tempopb"
 	"github.com/grafana/tempo/pkg/util/test"
 	"github.com/grafana/tempo/tempodb/backend"
@@ -165,6 +169,100 @@ func runnerTagValuesV2ClientCancelContext(t *testing.T, f *QueryFrontend) {
 	}
 	err := f.streamingTagValuesV2(grpcReq, srv)
 	require.Equal(t, status.Error(codes.Canceled, "context canceled"), err)
+}
+
+func TestSearchTagsV2Intrinsics(t *testing.T) {
+	mockScope := "span"
+	mockTags := []string{"foo", "bar"}
+
+	tcs := []struct {
+		name        string
+		maxTagBytes int
+		expected    *tempopb.SearchTagsV2Response
+	}{
+		{
+			name:        "unlimited",
+			maxTagBytes: 0,
+			expected: &tempopb.SearchTagsV2Response{
+				Scopes: []*tempopb.SearchTagsV2Scope{
+					{
+						Name: api.ParamScopeIntrinsic,
+						Tags: search.GetVirtualIntrinsicValues(),
+					},
+					{
+						Name: mockScope,
+						Tags: mockTags,
+					},
+				},
+			},
+		},
+		{
+			name:        "when_limited_intrinsics_first",
+			maxTagBytes: 100,
+			expected: &tempopb.SearchTagsV2Response{
+				Scopes: []*tempopb.SearchTagsV2Scope{
+					{
+						// Only a subset of intrinsic tags will fit
+						Name: api.ParamScopeIntrinsic,
+						Tags: search.GetVirtualIntrinsicValues()[0:10],
+					},
+				},
+			},
+		},
+	}
+
+	for _, tc := range tcs {
+		t.Run(tc.name, func(t *testing.T) {
+			// This is the mocked data returned by querier/ingester jobs downstream.
+			next := &mockRoundTripper{
+				responseFn: func() proto.Message {
+					return &tempopb.SearchTagsV2Response{
+						Scopes: []*tempopb.SearchTagsV2Scope{
+							{
+								Name: mockScope,
+								Tags: mockTags,
+							},
+						},
+					}
+				},
+			}
+
+			f := frontendWithSettings(t, next, nil, nil, nil, func(_ *Config, overridesCfg *overrides.Config) {
+				overridesCfg.Defaults.Read.MaxBytesPerTagValuesQuery = tc.maxTagBytes
+			})
+
+			// http
+			httpReq := httptest.NewRequest("GET", "/api/v2/search/tags", nil)
+			httpResp := httptest.NewRecorder()
+
+			ctx, cancel := context.WithCancel(httpReq.Context())
+			defer cancel()
+			ctx = user.InjectOrgID(ctx, "tenant")
+			httpReq = httpReq.WithContext(ctx)
+
+			f.SearchTagsV2Handler.ServeHTTP(httpResp, httpReq)
+			require.Equal(t, http.StatusOK, httpResp.Code)
+
+			resp := &tempopb.SearchTagsV2Response{}
+			bytesResp, err := io.ReadAll(httpResp.Body)
+			require.NoError(t, err)
+			err = jsonpb.Unmarshal(bytes.NewReader(bytesResp), resp)
+
+			require.NoError(t, err)
+
+			// Sort scopes to give stable comparison
+			sort.Slice(tc.expected.Scopes, func(i, j int) bool {
+				return tc.expected.Scopes[i].Name < tc.expected.Scopes[j].Name
+			})
+			sort.Slice(resp.Scopes, func(i, j int) bool {
+				return resp.Scopes[i].Name < resp.Scopes[j].Name
+			})
+			require.Equal(t, len(tc.expected.Scopes), len(resp.Scopes))
+			for i := range tc.expected.Scopes {
+				require.ElementsMatch(t, tc.expected.Scopes[i].Tags, resp.Scopes[i].Tags)
+			}
+		})
+	}
 }
 
 // todo: a lot of code is replicated between all of these "failure propagates from queriers" tests. we should refactor
@@ -477,7 +575,9 @@ func TestSearchTagsV2AccessesCache(t *testing.T) {
 	hash := fnv1a.HashString64(scope)
 	start := uint32(10)
 	end := uint32(20)
-	cacheKey := cacheKey(cacheKeyPrefixSearchTag, tenant, hash, int64(start), int64(end), meta, 0, 1)
+	startTime := time.Unix(int64(start), 0)
+	endTime := time.Unix(int64(end), 0)
+	cacheKey := cacheKey(cacheKeyPrefixSearchTag, tenant, hash, startTime, endTime, meta, 0, 1)
 
 	// confirm cache key coesn't exist
 	_, bufs, _ := c.Fetch(context.Background(), []string{cacheKey})